+++ prototyper/prototyper/src/macros.rs
@@ -27,14 +27,14 @@ macro_rules! println {
 macro_rules! has_csr {
     ($($x: expr)*) => {{
             use core::arch::asm;
-            use riscv::register::mtvec;
+            use ::riscv::register::mtvec;
             use crate::sbi::early_trap::light_expected_trap;
             let res: usize;
             unsafe {
                 // Backup old mtvec
                 let mtvec = mtvec::read().bits();
                 // Write expected_trap
-                mtvec::write(light_expected_trap as _, mtvec::TrapMode::Direct);
+                mtvec::write(mtvec::Mtvec::new(light_expected_trap as *const () as _, mtvec::TrapMode::Direct));
                 asm!("addi a0, zero, 0",
                     "addi a1, zero, 0",
                     "csrr a2, {}",