+++ library/sbi-rt/CHANGELOG.md
@@ -9,6 +9,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+### Modified
+
+### Fixed
+
+## [0.0.4] - 2026-05-13
+
+### Added
+
 - pmu: add missing `pmu_snapshot_set_shmem` function.
 - pmu: `pmu_snapshot_set_shmem` function signature, documents and implementation
 - lib: re-export `sbi_spec::base::CounterMask` on crate root.
@@ -20,6 +28,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - rt: add SSE extension support to SBI implementation.
 - feat(rt): add MPXY extension support to SBI runtime library.
 - lib: add `raw` module for raw SBI calls; they may be used in runtime libraries for custom SBI extensions.
+- lib: add raw RISC-V SBI calls with 3 to 6 parameters.
 
 ### Modified