+++ prototyper/CHANGELOG.md
@@ -7,7 +7,10 @@ All notable changes to this project will be documented in this file. See [conven
 ## Unreleased
 
 ### Added
+- Add AIA IMSIC IPI backend support for RustSBI Prototyper.
 
 ### Modified
-
+- Refine CSR group comments.
+- fix(prototyper): temporary PMU fix for possible S-mode DTB modification
+- fix(prototyper): validate DBCN console shared memory range
 ### Removed